+++ packages/common-ethereum/src/codegen/constants.ts
@@ -1,0 +1,9 @@
+// Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
+// SPDX-License-Identifier: GPL-3.0
+
+export const CUSTOM_EVM_HANDLERS = [
+  'cosmos/EthermintEvm',
+  'substrate/FrontierEvm',
+  'substrate/AcalaEvm',
+  'substrate/Moonbeam',
+];