+++ packages/common-ethereum/src/codegen/index.ts
@@ -1,5 +1,5 @@
 // Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
-export * from './codegen';
-export * from './project';
+export * from './codegen-controller';
+export * from './utils';