+++ packages/common-ethereum/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Migrated abi-codegen from `@subql/cli` to `common-ethereum` (#158)
 
 ## [2.2.5] - 2023-09-04
 ### Fixed